+++ get-user-custom-field-values.php
@@ -2,9 +2,9 @@
 /**
  * Plugin Name: Get User Custom Field Values
  * Version:     3.2
- * Plugin URI:  http://coffee2code.com/wp-plugins/get-user-custom-field-values/
+ * Plugin URI:  https://coffee2code.com/wp-plugins/get-user-custom-field-values/
  * Author:      Scott Reilly
- * Author URI:  http://coffee2code.com/
+ * Author URI:  https://coffee2code.com/
  * Text Domain: get-user-custom-field-values
  * License:     GPLv2 or later
  * License URI: https://www.gnu.org/licenses/gpl-2.0.html