+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## _(in-progress)_
 * Change: Update shortcode builder widget:
     * New: Store object instantiated during `register()`
+    * Change: Cast return value of `c2c_get_user_custom_field_values_post_types` filter as an array
 * New: Add TODO.md and move existing TODO list from top of main plugin file into it (and added to it)
 * Change: Note compatibility through WP 5.4+
 * Change: Update links to coffee2code.com to be HTTPS