+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## _(in-progress)_
-* Change: Update shortcode builder widget:
+* Change: Update shortcode builder widget to 007:
     * New: Store object instantiated during `register()`
     * Change: Cast return value of `c2c_get_user_custom_field_values_post_types` filter as an array
     * Change: Sanitize string used in markup attributes (hardening)